Implement the Meta module (#7616)
PR checklist:

- [x] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [ ] Tests included or PR comment includes a reproducible test plan
- [x] Documentation is up-to-date
- [ ] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [x] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)

This PR introduces the new module `Meta.ml` which allows `osemgrep` to
aggregate metadata from environment. Several questions persists and I
reach the point where the code seems correct. However, the PR is not
finished where I implemented only Git and GitHub modules.

These are the questions that arise and may have an impact on the
continuity of this work:
1) Dependencies, currently I added 3 dependencies:
- `bos` to invoke Git and get some informations about the Git repository
- [`digestif`](https://github.com/mirage/digestif) to have a clean
representation of the SHA1
- [`emile`](https://github.com/dinosaure/emile) to have something which
represents an email address (correctly)

Only `bos` is highly required even if we can replace it by something
else. About SHA1 and email address, we can keep a simple `string` to
represent them.

2) the `Glom` module. It's a poor lense module to get metadata from a
JSON file available into the CI environment. This module is only used to
implement the `GitHub` module. I'm not sure that something apparent
already exists so I implemented something quick and efficient for the
purpose. The main point is that it uses a GADT to "coerce" JSON types
into caml types (which can increase the complexity and understanding of
this part of the code).

3) The design of `Meta.ml` in general. `meta.py` is built around a base
class which the other classes must inherit. This kind of mechanism
exists in OCaml, but the use of objects and classes is not common with
this language. I decided to start with a fairly simple model of a
`module type S` implementing the properties expected of classes in
Python and which the various implementations should respect.

4) What seems to be most interesting is the "compilation" of the context
into a Python dictionary (the `to_dict` method). I have reproduced this
method in "compiling" to a "record" `type 'extension t` which can be
extended (I have indeed noticed that the dictionary can have more
metadata than initially expected).

5) The last point concerns the collection of information. I've tried to
detach the calculation of the various metadata from what should be done
when we run the `Cmdliner.Term.eval`. Thus, I expose "cmdliner" terms so
that they can be reused (and especially composed with what will happen
to `cli_ci/Ci_CLI.ml`). I could have just used `Sys.getenv` but I prefer
not to base my work on any global state.

For these different points, this is the design that should follow for
the other contexts (CircleCI, Bitbucket, etc.). In order to avoid
wasting time, I would like to clarify these points before continuing
further in order to start moving in the right direction :+1:.

---------

Co-authored-by: Yoann Padioleau <yoann.padioleau@gmail.com>