feat(js)!: unix.js (#9090)
## What:
This PR implements the Unix syscall dependencies of Semgrep in
Javascript, so that it may be used by the JSCaml code.

## Why:
When transpiling the language server code in OCaml to Javascript, we
must also be able to have the Javascript code invoke the OCaml
dependencies. Namely, this takes the form of the basis library's Unix
dependencies, such as syscalls like `caml_unix_spawn` and friends, which
are normally external C calls.

Since we're aiming to run everything in Javasceript, we can't use those
C calls, however. So we re-implement them to run in JS.

## How:
Re-implemented the calls in JS. Also did some organization.

## Test plan:
`make test` from `js/tests`.