Add 'semgrep scan --x-ls' subcommand and git-based listing in osemgrep (#9245)
osemgrep now calls `git ls-files` if a git project was detected.

I still don't have unit tests for this but if the CI tests pass, we
should probably merge it to minimize conflicts.

test plan: from within a repo such as the semgrep repo:
```
$ semgrep scan --x-ls | tee ls-py.out | grep '^+' > ls-selected-py.out
$ semgrep scan --x-ls --experimental | tee ls-ml.out | grep '^+' > ls-selected-ml.out
$ diff -u ls-selected-py.out ls-selected-ml.out | less
```

The original output of `--x-ls` (files `ls-py.out` and `ls-ml.out`)
shows the files that were selected as well as those that are ignored,
with the reason why they're being ignored. The files
`ls-selected-py.out` and `ls-selected-ml.out` contain only the selected
targets and the diff command shows the differences between pysemgrep and
osemgrep. They provide different results and I need to investigate
further.

Update: osemgrep no longer ignores dotfiles automatically. 3 tests fail
apparently because a `.git` folder isn't being ignored. I'm not sure
what mechanism causes it to be ignored in pysemgrep when the current
project is _not_ a git project. I'll look into this tomorrow.