refactor(output): port autofix to core (#9229)
## What:
This PR ports both metavariable interpolation and autofix to the core
Semgrep engine, deleting (some of) the `pysemgrep` and `osemgrep` code
for both.

## Why:
It's kind of gross that this logic existed only in the CLI side in the
first place. Furthermore, this led to a disparity between the core and
CLI outputs, where AST-based autofix would be applied in the first, but
not textual autofix. This is conceptually hazardous.

## How:
This was somewhat involved, due to the testy nature of the dependencies
of all of the data needed to make this change happen. The primary forces
at work here were due to the facts that AST-based autofix needs to
_consume_ the actual tree of the fix (thus, the `Pattern_match.t`), and
it needs to be able to _produce_ the fix information for the ATD semgrep
output, which needs to be able to associate each match with its
replacement text.

This PR introduces a new processor called `Autofix_processor`, whose
goal is to instrument a core scan by populating its autofix information.
Notably, now `Core_result` has the `matches_with_fixes` field, which is
not just `Pattern_match.t`, but tuples of `Pattern_match.t * Textedit.t
option`.

Since this PR aims to make autofix a functionality of the core engine,
and the `Core_json_output` module must be able to turn a `Core_result.t`
into an `Out.core_output`, we need this data to exist somewhere, so it
makes sense to me that it should become part of `Core_result.t`

## Alternatives considered:

### Why not put it at the `Pattern_match.t` -> `Pattern_match.t` level?
I did not want to add the fix information into the `Pattern_match.t`
type. That felt too far forward.

### Why not put it at the `Core_json_output` and `match_to_match` level?
I felt that this would conflate the idea of producing the JSON output
and producing the autofixes themselves. This also felt too "interleaved"
-- I wanted a horizontal step of producing _all_ of the
`Out.core_match`es, and a _separate_ step which produces and applies the
autofixes, across each match, as opposed to a vertical step where we
produce a single `Out.core_match` and then autofix it.

This also has dreadful consequences because the `Autofix` module is not
available from `Core_json_output`. You would have to pass in a
functional argument or something, plus the `autofix` flag.

### Why keep the Python code for `autofix.py` and not apply fixes in the
core engine?
That was the original plan. If you apply the autofixes in the core
engine, however, this will cause the match report to be wrong, as the
CLI will report the matches of the edited file. Thus, the autofixes need
to occur after that point.


## Test plan:
Automated tests.

I verified that the new tests fail when they are mangled.

---------

Co-authored-by: pad <pad@r2c.dev>