Mass renaming: Alcotest_ext -> Testo (#9589)
Alcotest_ext is now known as [Testo and has its own
repo](https://github.com/semgrep/testo). It is now a git submodule
checked out as `libs/testo`. I used the opportunity to remove the legacy
support for tests specified as a tuple (name, func). All tests in
semgrep are now created directly with a call to `Testo.create` which
offers options that a tuple can't provide.

Besides the renaming of the module `Alcotest_ext` to `Testo`, the
following changed:

* `pack_tests`, which used to operate on tuples (name, func), is gone.
* `pack_tests_pro` which works on the new test type is now known as
`Testo.categorize`.
* `pack_suites` is now known as `Testo.categorize_suites`.

The functions above were renamed to allude to the `category` field that
they modify.

I removed the few places where we had `open Testo` and added an alias
`let t = Testo.create` in all test modules.

At the time of writing, the work is done for the semgrep repo but needs
to be done for semgrep-proprietary.