experiment: use opentelemetry as a tracing collector (#9561)
Add two simple traces that can be activated via `--trace`. When
`--trace` isn't passed, the trace calls should be virtually no-ops.
When it is, spans for the trace will be sent to the collector.

Currently, this collector is just `localhost:4317`. Later, we'll
make it a permanent endpoint and include traces in metrics.

Test plan:
See instructions in notion "How to add tracing" to get this running

Non-blocking TODOs

TODO: does not yet work with multiple processes. I believe that to make
this work we're going to need to
start the trace in each process and attach the spans to a parent span.
([ticket](https://linear.app/semgrep/issue/EA-325/make-spans-work-with-multiple-processes))

TODO: instrument with the official ppx instead
([ticket](https://linear.app/semgrep/issue/EA-850/instrument-with-official-ppx))

TODO: add tags to the top level span for categorization
([ticket](https://linear.app/semgrep/issue/EA-324/tag-top-level-span-with-product-version-etc))

---------

Co-authored-by: Tom Petr <tom@r2c.dev>
Co-authored-by: ajbt200128 <austin@cutedogs.org>