feat: track content origins independent of files (#9715)
Currently we use a combination of `Fpath.t` and `string` pervasively for
a couple distinct purposes:
- reporting where a finding originated and
- getting the contents of the object we’re scanning (e.g., with
UFile.read_file).

This dual-purpose use means it is difficult to:
- create a target where the contents are not the verbatim file contents
- associate additional data to findings about location which we might
want to report to users(e.g., “nested within a metavariable pattern”;
“via extract mode”)
- scan anything which is not resident on the filesystem (e.g.,
metavariable pattern, git object)---currently this *requires* writing a
tempfile.

Mainly, this PR adds one additional type, ~`Source.t`~ `Origin.t`, which
acts as a "user facing" location we can report. In order to utilize
this, we also add several additional types (See ~`Target_location`~
`Target`) which serve to replace `Input_to_core.t` and similar inside of
the engine. Conversion from prior types to these new types is added
where needed. The types in ~`Target_location`~ `Target` then have a
field ~`source : Source.t`~ `origin : Origin.t` to attach the additional
information.

In this PR, ~`Source.t`~ `Origin.t` only has a single variant, but I
would anticipate adding more in the future (for instance, `GitBlob` for
historical scanning).

Additionally, a file path is currently present in these types for the
existing pattern of writing data to a tempfile for the purpose of
scanning. However, I think we can eventually get rid of the file path
for instances where there is not an advantage to writing the content to
disk, but I don't do so here.

See also discussion in
https://semgrepinc.slack.com/archives/C05STGMM53Q/p1706034545394119

---

For reviewers: I think this is best reviewed commit-by-commit; I have
tried to put up a rather clean history.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>