Remove credentials from project URL if present before returning (#9849)
`semgrep publish` uses `get_project_url` to get the remote URL for the
rule source and add it to the metadata for the rule. GitLab CI puts job
tokens in the remote URL, which means those tokens can be propagated to
the rule metadata if rules are published from GL CI/CD, which is not
desirable. With this change, credentials in the project URL are stripped
before it's returned.

Project URL handling for `semgrep ci` in GitLab CI/CD is done
differently and will not be affected by this change.

I've added a basic test, but I'm happy to add more or different tests if
what I added isn't adequate.