Mask also temporary folder obtained after resolving the symlinks (#10018)
This is a workaround to the issue of temporary folders being symlinks as
described in
https://github.com/semgrep/testo/pull/42.

This is not done in Testo because I find it too restrictive for Testo to
require OCaml >= 4.13 at this time.

The other bug addressed by https://github.com/semgrep/testo/pull/42
(trailing slashes in temp folder path) will be fixed in Testo.

Test plan: CI checks must pass but symlinked temp folders still fail for
other reasons:
```
$ ln -s /tmp link-to-tmp
$ TMPDIR=$(pwd)/link-to-tmp make test
```
1. If the temp folder is relative, the tests fail because of a chdir.
It's understandable.
2. With an absolute path such as `/home/martin/link-to-tmp`, 12 tests
fail. Here's the error we get for test 1d84c9d09005 (`Language Server
(unit) > Session Targets > Test multiple workspaces with some dirty
(only_git_dirty: false)`):
```
 FAIL: The test raised an exception: Failure: cannot make path "/home/martin/semgrep2/link-to-tmp/test_workspace_9c93dc2d-def6-48f2-84cc-65eeade3e0d2" relative to project root "/tmp/test_workspace_9c93dc2d-def6-48f2-84cc-65eeade3e0d2".
 cwd: /home/martin/semgrep2
 realpath for .: { Rfpath.fpath = .; rpath = (Rpath.Rpath /home/martin/semgrep2);
   cwd = (Rpath.Rpath /home/martin/semgrep2) }
 Sys.argv: ./test
```
It is a known bug that is not simple to fix. See
https://github.com/semgrep/semgrep/pull/10003

Some other language server tests fail with 30-second timeouts. This
should be fixed such that exceptions cause an immediate failure.